% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bearing_polarization.R
\name{bearing_polarization}
\alias{bearing_polarization}
\title{Calculate group polarization}
\usage{
bearing_polarization(DT, bearing = "bearing", group = "group", degree = FALSE)
}
\arguments{
\item{DT}{with bearing column generated by eg. bearing_sequential}

\item{bearing}{bearing column name, default 'bearing'}

\item{group}{group column name generated by eg. group_pts, default 'group'}

\item{degree}{boolean, if bearing is measured in degrees. See ?CircStats::r.test}
}
\description{
Polarization of group bearings from 0-1
}
