% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bearing_sequential.R
\name{bearing_sequential}
\alias{bearing_sequential}
\title{Calculate absolute sequential bearing}
\usage{
bearing_sequential(DT, id = NULL, coords = NULL, projection = NULL)
}
\description{
Ensure input DT is ordered by datetime or timegroup using
eg. setorder(DT, timegroup)
}
