% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bearing_to_group_centroid.R
\name{bearing_to_group_centroid}
\alias{bearing_to_group_centroid}
\title{Calculate absolute bearing to group centroid}
\usage{
bearing_to_group_centroid(DT, coords = NULL)
}
\arguments{
\item{DT}{expects group_mean columns generated with group_centroid}

\item{coords}{character vector of column names for x, y}
}
\description{
Calculate absolute bearing to group centroid
}
