% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bearing_to_leader.R
\name{bearing_to_leader}
\alias{bearing_to_leader}
\title{Calculate absolute bearing to group leader}
\usage{
bearing_to_leader(DT, coords = c("x", "y"), group = "group")
}
\arguments{
\item{coords}{character vector columns names for x and y}

\item{group}{group column name generated by eg. group_pts}
}
\description{
Leader as identified using group_leader by first position along
group axis of movement (return_rank = TRUE)
}
\details{
Note: coords must be in order x, y
}
