% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build-lines.R
\name{build_lines}
\alias{build_lines}
\title{Build Lines}
\usage{
build_lines(dt, crs, coord.fields = c("EASTING", "NORTHING"),
  id.field = "ID")
}
\arguments{
\item{dt}{Input data with coordinate and id fields. If not provided, fields default to c('EASTING', 'NORTHING') and 'ID'}

\item{crs}{Character string for input to sp::CRS()}

\item{coord.fields}{Character vector indicating the X coordinate and Y
coordinate columns. eg: c('EASTING', 'NORTHING')}

\item{id.field}{Character string indicating the column name for the ID associated
with each point}
}
\value{
SpatialLines for each ID provided
}
\description{
Build Lines
}
\examples{

}
