% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_lines_sp.R
\name{build_lines_sp}
\alias{build_lines_sp}
\title{Build Lines (deprecated version of function with retired spatial packages)}
\usage{
build_lines_sp(
  DT = NULL,
  projection = NULL,
  id = NULL,
  coords = NULL,
  sortBy = NULL,
  splitBy = NULL
)
}
\arguments{
\item{DT}{input data.table}

\item{projection}{numeric or character defining the coordinate reference
system to be passed to \link[sf:st_crs]{sf::st_crs}. For example, either
\code{projection = "EPSG:32736"} or \code{projection = 32736}.}

\item{id}{Character string of ID column name}

\item{coords}{Character vector of X coordinate and Y coordinate column names}

\item{sortBy}{Character string of date time column(s) to sort rows by. Must
be a POSIXct.}

\item{splitBy}{(optional) character string or vector of grouping column
name(s) upon which the grouping will be calculated}
}
\description{
Build Lines (deprecated version of function with retired spatial packages)
}
\seealso{
\code{\link{group_lines}}

Other Build functions: 
\code{\link{build_lines}()},
\code{\link{build_polys_sp}()},
\code{\link{build_polys}()}
}
\concept{Build functions}
