% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_polys.R
\name{build_polys}
\alias{build_polys}
\title{Build Home Ranges for Individuals}
\usage{
build_polys(DT = NULL, projection = NULL, hrType = NULL,
  hrParams = NULL, coords = NULL, id = NULL, splitBy = NULL,
  spPts = NULL)
}
\arguments{
\item{DT}{input data.table}

\item{projection}{PROJ.4 character string}

\item{hrType}{type of HR estimation, of 'mcp' or 'kernel'}

\item{hrParams}{parameters for adehabitatHR functions, a named list passed to do.call}

\item{coords}{Character vector of X coordinate and Y coordinate column names}

\item{id}{Character string of ID column name}

\item{splitBy}{(optional) character string or vector of grouping field(s) upon which the grouping will be calculated}

\item{spPts}{alternatively, provide solely a SpatialPointsDataFrame with one column representing the ID of each point.}
}
\value{
Home range polygons for each ID
}
\description{
Build Home Ranges for Individuals
}
