% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_polys_sp.R
\name{build_polys_sp}
\alias{build_polys_sp}
\title{Build Polygons (deprecated version of function with retired spatial packages)}
\usage{
build_polys_sp(
  DT = NULL,
  projection = NULL,
  hrType = NULL,
  hrParams = NULL,
  id = NULL,
  coords = NULL,
  splitBy = NULL,
  spPts = NULL
)
}
\arguments{
\item{DT}{input data.table}

\item{projection}{numeric or character defining the coordinate reference
system to be passed to \link[sf:st_crs]{sf::st_crs}. For example, either
\code{projection = "EPSG:32736"} or \code{projection = 32736}.}

\item{hrType}{type of HR estimation, either 'mcp' or 'kernel'}

\item{hrParams}{a named list of parameters for \code{adehabitatHR} functions}

\item{id}{Character string of ID column name}

\item{coords}{Character vector of X coordinate and Y coordinate column names}

\item{splitBy}{(optional) character string or vector of grouping column
name(s) upon which the grouping will be calculated}

\item{spPts}{alternatively, provide solely a SpatialPointsDataFrame with one
column representing the ID of each point, as specified by \link[adehabitatHR:mcp]{adehabitatHR::mcp}
or \link[adehabitatHR:kernelUD]{adehabitatHR::kernelUD}}
}
\description{
Build Polygons (deprecated version of function with retired spatial packages)
}
\seealso{
\code{\link{group_polys}}

Other Build functions: 
\code{\link{build_lines_sp}()},
\code{\link{build_lines}()},
\code{\link{build_polys}()}
}
\concept{Build functions}
