% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build-lines.R
\name{BuildLines}
\alias{BuildLines}
\title{Build Lines}
\usage{
BuildLines(DT = NULL, projection = NULL, idField = NULL,
  coordFields = NULL, byFields = NULL)
}
\arguments{
\item{DT}{Input data.table with coordinate and id fields. If not provided, fields
default to c('EASTING', 'NORTHING') and 'ID'.}

\item{projection}{Character string for input to sp::CRS() and proj4string().}

\item{idField}{Character string indicating the column name for the ID
associated with each point.}

\item{coordFields}{Character vector indicating the X coordinate and Y
coordinate columns. eg: c('EASTING', 'NORTHING')}
}
\value{
SpatialLines for each ID provided
}
\description{
Build Lines
}
