% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build-pts.R
\name{BuildPts}
\alias{BuildPts}
\title{Build Points}
\usage{
BuildPts(DT, projection, coordFields = c("EASTING", "NORTHING"),
  idField = "ID")
}
\arguments{
\item{DT}{Input data with coordinate and id fields. If not provided, fields
default to c('EASTING', 'NORTHING') and 'ID'}

\item{projection}{Character string for input to sp::CRS() and proj4string()}

\item{coordFields}{Character vector indicating the X coordinate and Y
coordinate columns. eg: c('EASTING', 'NORTHING')}

\item{idField}{Character string indicating the column name for the ID
associated with each point}
}
\value{
SpatialPointsDataFrame for each ID provided
}
\description{
Build a SpatialPointsDataFrame from provided data for input to GroupPts or
other uses.
}
\details{
The parameter projection can be built by finding the EPSG code for your
given projection at \url{http://spatialreference.org/ref/epsg/} and providing it with
"+init=epsg:CODE" (eg: "+init=epsg:4326").
}
