% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{calc_centroid}
\alias{calc_centroid}
\title{Calculate centroid}
\usage{
calc_centroid(geometry, x, y, crs)
}
\arguments{
\item{geometry}{sfc (simple feature geometry list column) from \code{\link[=get_geometry]{get_geometry()}}}

\item{x}{X coordinate column, numeric}

\item{y}{Y coordinate column, numeric}

\item{crs}{crs for x, y coordinates, ignored for geometry argument}
}
\value{
Centroid of the geometry or coordinates in x,y provided
}
\description{
\strong{Internal function} - not developed to be used outside of spatsoc functions
}
\details{
Calculate centroid using \code{\link[sf:geos_unary]{sf::st_centroid()}} for one of:
\itemize{
\item geometry
\item the points in x, y
\item the pairwise points in geometry_a and geometry_b
\item the pairwise points in x_a, y_a and x_b, y_b
}
}
\examples{
# Load data.table
library(data.table)
\dontshow{
data.table::setDTthreads(1)
}
# Read example data
DT <- fread(system.file("extdata", "DT.csv", package = "spatsoc"))

DT[, spatsoc:::calc_centroid(x = X, y = Y, crs = 32736)]
}
\keyword{internal}
