% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/centroids-group.R
\name{CalcCentroids}
\alias{CalcCentroids}
\title{Calculate Centroids For Each Group}
\usage{
CalcCentroids(DT, coordFields, groupField)
}
\arguments{
\item{DT}{input data.table}

\item{coordFields}{Character vector of X coordinate and Y coordinate column names}

\item{groupField}{The group field generated}
}
\description{
Append two columns indicating the centroid coordinates for each group to an
group by ID, coordinates data.table/frame.
}
