% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/centroid_group.R
\name{centroid_group}
\alias{centroid_group}
\title{Group centroid}
\usage{
centroid_group(DT = NULL, coords = NULL, group = "group", na.rm = FALSE)
}
\arguments{
\item{DT}{input data.table with group column generated with \code{group_pts}}

\item{coords}{character vector of X coordinate and Y coordinate column names.
Note: the order is assumed X followed by Y column names.}

\item{group}{Character string of group column}

\item{na.rm}{if NAs should be removed in calculating mean location,
see \code{mean}}
}
\value{
\code{centroid_group} returns the input \code{DT} appended with
centroid columns for the X and Y coordinate columns.

These columns represents the centroid coordinate columns.
The naming of these columns will correspond to the provided coordinate
column names prefixed with "centroid_".

A message is returned when centroid columns are already exists in
the input \code{DT}, because they will be overwritten.

See details for appending outputs using modify-by-reference in the
\href{https://docs.ropensci.org/spatsoc/articles/faq.html}{FAQ}.
}
\description{
\code{centroid_group} calculates the centroid (mean location) of all
individuals in each spatiotemporal group identified by \code{group_pts}. The
function expects a \code{data.table} with relocation data appended with a
\code{group} column from \code{group_pts}. Relocation data should be in two
columns representing the X and Y coordinates.
}
\details{
The \code{DT} must be a \code{data.table}. If your data is a
\code{data.frame}, you can convert it by reference using
\code{\link[data.table:setDT]{data.table::setDT()}} or by reassigning using
\code{\link[data.table:data.table]{data.table::data.table()}}.

The \code{coords} and \code{group} arguments expect the names of a column in
\code{DT} which correspond to the X and Y coordinates and group columns. The
\code{na.rm} argument is passed to the \code{mean} function to control if NA
values are removed before calculation.
}
\examples{
# Load data.table
library(data.table)
\dontshow{data.table::setDTthreads(1)}

# Read example data
DT <- fread(system.file("extdata", "DT.csv", package = "spatsoc"))

# Cast the character column to POSIXct
DT[, datetime := as.POSIXct(datetime, tz = 'UTC')]

# Temporal grouping
group_times(DT, datetime = 'datetime', threshold = '20 minutes')

# Spatial grouping with timegroup
group_pts(DT, threshold = 5, id = 'ID',
          coords = c('X', 'Y'), timegroup = 'timegroup')

# Calculate group centroid
centroid_group(DT, coords = c('X', 'Y'), group = 'group', na.rm = TRUE)
}
\seealso{
\code{group_pts}

Other Centroid functions: 
\code{\link{centroid_dyad}()},
\code{\link{centroid_fusion}()},
\code{\link{direction_to_centroid}()},
\code{\link{distance_to_centroid}()}
}
\concept{Centroid functions}
