% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance_to_group_centroid.R
\name{distance_to_group_centroid}
\alias{distance_to_group_centroid}
\title{Calculate distance to group centroid}
\usage{
distance_to_group_centroid(DT, coords, group = "group", return_rank = FALSE)
}
\arguments{
\item{DT}{expects group_mean columns generated with group_centroid}

\item{coords}{character vector of column names for x, y}
}
\description{
Calculate distance to group centroid
}
