% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance_to_leader.R
\name{distance_to_leader}
\alias{distance_to_leader}
\title{Calculate distance to leader}
\usage{
distance_to_leader(DT, coords = c("x", "y"), group = "group")
}
\arguments{
\item{coords}{character vector of column names for x, y}

\item{group}{group column name generated by eg. group_pts}
}
\description{
Leader as identified using group_leader by first position along
group axis of movement (return_rank = TRUE)
}
