% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{DT}
\alias{DT}
\title{Example relocation data}
\format{A data.table with 4 columns: ID, x, y, datetime}
\source{
to fill...
}
\usage{
DT
}
\description{
An example dataset of three individuals' relocations
}
\keyword{datasets}
