% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edge_bearing.R
\name{edge_bearing}
\alias{edge_bearing}
\title{Bearing difference based edge lists}
\usage{
edge_bearing(
  DT = NULL,
  threshold,
  id = NULL,
  coords = NULL,
  timegroup,
  splitBy = NULL,
  returnDist = FALSE,
  fillNA = TRUE
)
}
\description{
Bearing difference based edge lists
}
\seealso{
Other Edge-list generation: 
\code{\link{edge_dist}()},
\code{\link{edge_nn}()}
}
\concept{Edge-list generation}
