% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edge_delay.R
\name{edge_delay}
\alias{edge_delay}
\title{Directional correlation delay edge lists}
\usage{
edge_delay(DT, id = NULL, edges, window = NULL)
}
\arguments{
\item{DT}{relocation data}

\item{edges}{edges generated with edges_dist}

\item{window}{integer window in timegroups generated with group_times}
}
\description{
Temporal delay in absolute bearing between individuals
}
