% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edge_dist.R
\name{edge_dist}
\alias{edge_dist}
\title{Distance based edge lists}
\usage{
edge_dist(DT = NULL, threshold = NULL, id = NULL, coords = NULL,
  timegroup = NULL, splitBy = NULL, fillNA = TRUE)
}
\arguments{
\item{DT}{input data.table}

\item{threshold}{distance for grouping points, in the units of the coordinates}

\item{id}{Character string of ID column name}

\item{coords}{Character vector of X coordinate and Y coordinate column names}

\item{timegroup}{(optional) timegroup field in the DT upon which the grouping will be calculated}

\item{splitBy}{(optional) character string or vector of grouping column name(s) upon which the grouping will be calculated}

\item{fillNA}{boolean indicating if NAs should be returned for individuals that were not within the threshold distance of any other. If TRUE, NAs are returned. If FALSE, only edges between individuals within the threshold distance are returned.}
}
\description{
\code{edge_dist} returns edge lists defined by a spatial distance within the user defined threshold. The function accepts a \code{data.table} with relocation data, individual identifiers and a threshold argument. The threshold argument is used to specify the criteria for distance between points which defines a group. Relocation data should be in two columns representing the X and Y coordinates.
}
\examples{
# Load data.table
library(data.table)

# Read example data
DT <- fread(system.file("extdata", "DT.csv", package = "spatsoc"))

# Cast the character column to POSIXct
DT[, datetime := as.POSIXct(datetime, tz = 'UTC')]

# Temporal grouping
group_times(DT, datetime = 'datetime', threshold = '20 minutes')

# Edge list generation
edge_dist(DT, threshold = 100, id = 'ID', coords = c('X', 'Y'), timegroup = 'timegroup')

}
\concept{Edge-list generation}
