% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fusion_id.R
\name{fusion_id}
\alias{fusion_id}
\title{Identify fusion events}
\usage{
fusion_id(
  edges,
  threshold = 50,
  n_min_length = 0,
  n_max_missing = 0,
  allow_split = FALSE
)
}
\arguments{
\item{edges}{edges generated by edge_dist}

\item{threshold}{spatial distance threshold in the units of the projection}

\item{n_min_length}{minimum length of fusion events}

\item{n_max_missing}{maximum number of missing observations within a fusion event}

\item{allow_split}{boolean if a single observation can be greater than the
threshold distance without initiating fission event}
}
\description{
Identify fusion events
}
