% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_bearing.R
\name{group_bearing}
\alias{group_bearing}
\title{Group mean bearing}
\usage{
group_bearing(DT, bearing = "bearing", group = "group")
}
\arguments{
\item{bearing}{bearing column name generated by eg. bearing_sequential,
default 'bearing'}

\item{group}{group column generated by eg. group_pts}
}
\description{
Group mean bearing
}
