% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_centroid.R
\name{group_centroid}
\alias{group_centroid}
\title{Group centroid}
\usage{
group_centroid(DT, coords, group = "group", na.rm = FALSE)
}
\arguments{
\item{DT}{expecting group generated with eg. group_pts}

\item{coords}{character vector of column names for x, y}

\item{group}{group column default 'group'}

\item{na.rm}{if NAs should be removed in calculating mean location}
}
\description{
Mean of individual locations in each group
}
