% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_dyad.R
\name{group_dyad}
\alias{group_dyad}
\title{Group Dyads}
\usage{
group_dyad(DT, n = 2, id1, id2, timegroup)
}
\arguments{
\item{DT}{input data.table}

\item{timegroup}{timegroup field in the DT upon which the grouping will be calculated}

\item{returnDist}{boolean indicating if the distance between individuals should be returned. If FALSE (default), only ID1, ID2 columns (and timegroup, splitBy columns if provided) are returned. If TRUE, another column "distance" is returned indicating the distance between ID1 and ID2.}

\item{fillNA}{boolean indicating if NAs should be returned for individuals that were not within the threshold distance of any other. If TRUE, NAs are returned. If FALSE, only edges between individuals within the threshold distance are returned.}
}
\description{
Group Dyads
}
\examples{
# Load data.table
library(data.table)

# Read example data
DT <- fread(system.file("extdata", "DT.csv", package = "spatsoc"))

# Cast the character column to POSIXct
DT[, datetime := as.POSIXct(datetime, tz = 'UTC')]

# Temporal grouping
group_times(DT, datetime = 'datetime', threshold = '20 minutes')


}
\seealso{
\code{\link{group_times}}

Other Spatial grouping: 
\code{\link{group_lines}()},
\code{\link{group_polys}()},
\code{\link{group_pts}()}
}
\concept{Spatial grouping}
