% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_leader.R
\name{group_leader}
\alias{group_leader}
\title{Position along group's mean bearing as leadership order}
\usage{
group_leader(
  DT,
  group_bearing = "group_mean_bearing",
  coords = c("x", "y"),
  group = "group",
  return_rank = FALSE,
  ties.method = "average"
)
}
\arguments{
\item{group_bearing}{group_bearing column name generated by eg. group_bearing,
default 'group_mean_bearing'}

\item{coords}{character vector of column names for x, y}

\item{group}{group column generated by eg. group_pts}

\item{return_rank}{if rank along mean bearing should also be returned,
used by functions bearing_to_leader, distance_to_leader}

\item{ties.method}{default: 'average', see ?data.table::frank. Note if
multiple individuals in a group have identical coordinates, they will
receive an average rank (eg. 1.5, 2.5) which may indicate some data cleaning
is required.}
}
\description{
Given the mean bearing of a group of animals, shifts the coordinate
system to a new origin at the group centroid and rotates the coordinate system
by the mean bearing. Returns the distance along the mean bearing as
the front-back position in group.
}
