% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_polys.R
\name{group_polys}
\alias{group_polys}
\title{Group Overlapping Polygons}
\usage{
group_polys(DT = NULL, area = NULL, hrType = NULL, hrParams = NULL,
  projection = NULL, coordFields = NULL, idField = NULL,
  groupFields = NULL, spPolys = NULL)
}
\arguments{
\item{DT}{input data.table}

\item{area}{boolean indicating either returning area of overlap
of polygons or simply group}

\item{hrType}{Type of HR estimation, defaults to 'mcp'}

\item{projection}{PROJ.4 charaster string}

\item{coordFields}{Character vector of X coordinate and Y coordinate column names}

\item{idField}{Character string of ID column name}

\item{groupFields}{(optional) character string or vector of grouping field(s) upon which the grouping will be calculated}

\item{spPolys}{Alternatively, provide a SpatialPolygons object.}
}
\value{
Group by ID (by time) data.table
}
\description{
Group individuals by polygon (eg: home ranges) overlap
}
\examples{
# Build the HRs for a set of locs by individuals using mcp,
# kernel density estimation...
data(locs)

groups <- group_polys('mcp', locs, 50, projection = '+proj=utm +zone=21 ellps=WGS84',
                   idField = 'ID')

# If you'd like to simply compare proportion or overlap of a set of polygons,
# ...
data(locsPolys)

groups <- group_polys(spPolys = locsPolys)
}
