% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_times.R
\name{group_times}
\alias{group_times}
\title{Group Times}
\usage{
group_times(DT = NULL, timeField = NULL, threshold = NULL)
}
\arguments{
\item{DT}{input data.table}

\item{timeField}{name of time column}

\item{threshold}{threshold for grouping times.
eg: '2 hours', '10 minutes', etc.
if not provided, times will be matched exactly.
Note that provided threshold must be in the expected format: '## unit'}
}
\description{
Assign an integer timegroup.
Provide the threshold to group on and rows will be grouped.
If the threshold is NULL, rows are grouped by matching exact datetimes.
}
\details{
The threshold can be in units of minutes, hours or days.
}
\examples{
group_times(locs, timeField = 'datetime', threshold = '5 minutes')

group_times(locs, timeField = 'datetime', threshold = '2 hours')

group_times(locs, timeField = 'datetime', threshold = '10 days')

}
