% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group-hrs.R
\name{GroupHRs}
\alias{GroupHRs}
\title{Group Overlapping Polygons (Home ranges)}
\usage{
GroupHRs(hrType = "mcp", dt, crs, coordFields = c("EASTING", "NORTHING"),
  idField = "ID", spPolys = NULL)
}
\arguments{
\item{hrType}{Type of HR estimation, defaults to 'mcp'}

\item{dt}{Input data with coordinate and id fields. If not provided, fields default to c('EASTING', 'NORTHING') and 'ID'}

\item{crs}{Character string for input to sp::CRS()}

\item{coordFields}{Character vector indicating the X coordinate and Y
coordinate columns. eg: c('EASTING', 'NORTHING')}

\item{idField}{Character string indicating the column name for the ID associated
with each point}

\item{spPolys}{Alternatively, provide a SpatialPolygons object.}
}
\value{
id by group data.table
}
\description{
Group individuals by polygon overlap
}
