% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group-pts.R
\name{GroupPts}
\alias{GroupPts}
\title{Group Points}
\usage{
GroupPts(DT = NULL, threshold = NULL, idField = NULL,
  coordFields = NULL, timeGroup = NULL, groupFields = NULL)
}
\arguments{
\item{DT}{input data.table}

\item{threshold}{for grouping points, in the units of the projection}

\item{idField}{Character string of ID column name}

\item{coordFields}{Character vector of X coordinate and Y coordinate column names}

\item{timeGroup}{(optional) timegroup field in the DT upon which the grouping will be calculated}

\item{groupFields}{(optional) character string or vector of grouping field(s) upon which the grouping will be calculated}
}
\value{
Input data.table with column 'group' added.
}
\description{
Group points by spatial and temporal overlap.
}
\details{
This function finds spatialtemporal groups in input points. The threshold provided must be in the units of the projection of the coordinates. UTM coordinates (recommended) are in meters and threshold = 50 indicates a 50m threshold.
}
\examples{

GroupPts(locs, threshold = 5, idField = 'ID',
         coordFields = c('X', 'Y'))

GroupPts(locs, threshold = 5, idField = 'ID',
         coordFields = c('X', 'Y'), timeGroup = 'timegroup')

GroupPts(locs, threshold = 5, idField = 'ID', coordFields = c('X', 'Y'),
         timeGroup = 'timegroup', groupFields = 'season')
}
