% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time-groups.R
\name{GroupTime}
\alias{GroupTime}
\title{Group Time}
\usage{
GroupTime(dt, timeField, roundField = NULL)
}
\arguments{
\item{dt}{input locs/rows}

\item{timeField}{time column name}

\item{roundBool}{(optional) unit of time to round on (eg: day, hour, minute)}
}
\description{
Assign a numerical group for rows by provided time column. If
using GPS collar data, it is recommended to round this time column providing
the rounding unit as described below.
}
\details{
The rounding unit must be provided

Note that this numerical time group ID is added to the provided data.table.
}
