% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group-time.R
\name{GroupTimes}
\alias{GroupTimes}
\title{Group Times}
\usage{
GroupTimes(DT, timeField, timeThreshold = NULL)
}
\arguments{
\item{DT}{input locs/rows}

\item{timeField}{time column name}

\item{timeThreshold}{threshold for grouping times. eg: '2 hours', '10 minutes', etc.
if not provided, times will be matched exactly. Note that provided
threshold must be in the expected format: ## unit}
}
\description{
Assign a numerical group for rows by provided time column. If
using GPS collar data, it is recommended to round this time column providing
the rounding unit as described below.
}
\details{
The rounding unit must be provided....

Note that this numerical time group ID is added to the provided data.table.
}
