% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groups-lines.R
\name{grp.lines}
\alias{grp.lines}
\title{Groups From Spatial Lines}
\usage{
grp.lines(in.dt, crs)
}
\arguments{
\item{in.dt}{the 3 column data.table with coordinates (EASTING, NORTHING) and
ID. if non-standard column names are used, provide them with arguments..}
}
\value{
id by spatial group data.table
}
\description{
Groups From Spatial Lines
}
\examples{

}
