% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groups-locs.R
\name{grp_pts}
\alias{grp_pts}
\title{Group Points By Buffer}
\usage{
grp_pts(sp.pts, width)
}
\arguments{
\item{sp.pts}{The SpatialPointsDataFrame as created by the sp package. The}

\item{width}{The width of the buffer in the units of the projection}
}
\description{
Group points by buffer overlap at each time interval
}
\details{
this function uses input spatial points to determine
}
