% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iterations.R
\name{Iterations}
\alias{Iterations}
\title{Iterations}
\usage{
Iterations(DT, idField, iterations, groupField, randomType, dateField = NULL)
}
\arguments{
\item{DT}{input data.table with id, group fields and (optional) time fields}

\item{idField}{field indicating the id in the input data.table}

\item{iterations}{The number of iterations to randomize}

\item{groupField}{field indicating the group membership for each id}

\item{randomType}{one of 'daily', 'hourly' or 'spiegel'. 'daily' will randomize the ID
for each individual 24hr trajector while 'hourly' simply randomly assigns
an ID to each location. 'spiegel' will implement the daily movement
trajectory randomizations (Spiegel et al. 2016).}
}
\description{
Iterations
}
\seealso{
\url{http://onlinelibrary.wiley.com/doi/10.1111/2041-210X.12553/full}
}
