% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean-pairwise-dist.R
\name{mean_pairwise_dist}
\alias{mean_pairwise_dist}
\title{Euclidean Pairwise Distance}
\usage{
mean_pairwise_dist(in.dt, time.col = NULL, id.col = "ID",
  east.col = "EASTING", north.col = "NORTHING")
}
\arguments{
\item{time.col}{time column upon which individuals are compared. recommended
to use a rounded time, see SOME LINK.}

\item{id.col}{ID column name, if not provided defaults to 'ID'}

\item{east.col}{X column name, if not provided defaults to 'EASTING'}

\item{north.col}{Y column name, if not provided defaults to 'NORTHING'}

\item{locs}{input data.table with X, Y and ID columns}
}
\description{
Euclidean Pairwise Distance
}
