% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/round-time.R
\name{RoundTime}
\alias{RoundTime}
\title{Round Time}
\usage{
RoundTime(DT, timeField, roundUnit = "1 hour", dateField)
}
\arguments{
\item{DT}{input locs/rows}

\item{timeField}{time column name}

\item{roundUnit}{unit of time to round on (eg: day, hour, minute)}
}
\description{
Round Time
}
