% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time-groups.R
\name{time_groups}
\alias{time_groups}
\title{Time Groups}
\usage{
time_groups(dt, time, roundunit = NULL)
}
\arguments{
\item{dt}{input locs/rows}

\item{time}{time column name}

\item{roundunit}{unit by which the time should be rounded (optional)}
}
\description{
Assign a numerical group ID for rows/locs sorted by provided time column. If
using GPS collar data, it is recommended to round this time column providing
the rounding unit as described below.
}
\details{
The rounding unit must be provided

Note that this numerical time group ID is added to the provided data.table.
}
