% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_lines.R
\name{build_lines}
\alias{build_lines}
\title{Build Lines}
\usage{
build_lines(DT = NULL, projection = NULL, id = NULL, coords = NULL,
  splitBy = NULL)
}
\arguments{
\item{DT}{input data.table}

\item{projection}{PROJ.4 character string}

\item{id}{Character string of ID column name}

\item{coords}{Character vector of X coordinate and Y coordinate column names}

\item{splitBy}{(optional) character string or vector of grouping field(s) upon which the grouping will be calculated}
}
\value{
SpatialLines for each ID provided
}
\description{
Build Lines
}
