% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build-hrs.R
\name{BuildHRs}
\alias{BuildHRs}
\title{Build Home Ranges for Individuals}
\usage{
BuildHRs(hrType = "mcp", dt, projection, coordFields = c("EASTING",
  "NORTHING"), idField = "ID", spPts = NULL)
}
\arguments{
\item{hrType}{Type of HR estimation, defaults to 'mcp'}

\item{dt}{Input data with coordinate and id fields. If not provided, fields
default to c('EASTING', 'NORTHING') and 'ID'}

\item{projection}{Character string for input to sp::CRS() and proj4string()}

\item{coordFields}{Character vector indicating the X coordinate and Y
coordinate columns. eg: c('EASTING', 'NORTHING')}

\item{idField}{Character string indicating the column name for the ID
associated with each point}
}
\value{
Home range polygons for each ID
}
\description{
Build Home Ranges for Individuals
}
