% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build-hrs.R
\name{BuildHRs}
\alias{BuildHRs}
\title{Build Home Ranges for Individuals}
\usage{
BuildHRs(DT = NULL, projection = NULL, hrType = NULL, hrParams = NULL,
  coordFields = NULL, idField = NULL, byFields = NULL, spPts = NULL)
}
\arguments{
\item{DT}{input data.table}

\item{projection}{PROJ.4 charaster string}

\item{hrType}{Type of HR estimation either 'mcp' or 'kernel'}

\item{hrParams}{List of named arguments to be passed to adehabitatHR functions as determined by hrType}

\item{coordFields}{Character vector of X coordinate and Y coordinate column names}

\item{idField}{Character string of ID column name}
}
\value{
Home range polygons for each ID
}
\description{
Build Home Ranges for Individuals
}
