% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build-lines.R
\name{build_lines}
\alias{build_lines}
\title{Build Lines}
\usage{
build_lines(DT = NULL, projection = NULL, idField = NULL,
  coordFields = NULL, byFields = NULL)
}
\arguments{
\item{DT}{input data.table}

\item{projection}{PROJ.4 charaster string}

\item{idField}{Character string of ID column name}

\item{coordFields}{Character vector of X coordinate and Y coordinate column names}
}
\value{
SpatialLines for each ID provided
}
\description{
Build Lines
}
