% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build-lines.R
\name{BuildLines}
\alias{BuildLines}
\title{Build Lines}
\usage{
BuildLines(dt, crs, coordFields = c("EASTING", "NORTHING"), idField = "ID")
}
\value{
SpatialLines for each ID provided
}
\description{
Build Lines
}
\examples{

}
