% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_pts.R
\name{BuildPts}
\alias{BuildPts}
\title{Build Points}
\usage{
BuildPts(DT, projection, coords = c("EASTING", "NORTHING"), id = "ID")
}
\arguments{
\item{DT}{input data.table}

\item{projection}{PROJ.4 character string}

\item{coords}{Character vector of X coordinate and Y coordinate column names}

\item{id}{Character string of ID column name}
}
\value{
SpatialPointsDataFrame
}
\description{
Build a SpatialPointsDataFrame from provided data.table.
}
\details{
The parameter projection can be built by finding the EPSG code for your
given projection at \url{http://spatialreference.org/ref/epsg/} and providing it with
"+init=epsg:CODE" (eg: "+init=epsg:4326").
}
