% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build-points.R
\name{BuildPts}
\alias{BuildPts}
\title{Build Points}
\usage{
BuildPts(dt, crs, coordFields = c("EASTING", "NORTHING"), idField = "ID")
}
\arguments{
\item{dt}{Input data with coordinate and id fields. If not provided, fields default to c('EASTING', 'NORTHING') and 'ID'}

\item{crs}{Character string for input to sp::CRS()}

\item{coordFields}{Character vector indicating the X coordinate and Y
coordinate columns. eg: c('EASTING', 'NORTHING')}

\item{idField}{Character string indicating the column name for the ID associated
with each point}
}
\value{
SpatialPointsDataFrame for each ID provided
}
\description{
Build a SpatialPointsDataFrame from provided data for input to GroupPts or other uses.
}
