% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{calc_distance}
\alias{calc_distance}
\title{Calculate distance}
\usage{
calc_distance(geometry_a, geometry_b, x_a, y_a, x_b, y_b, crs)
}
\arguments{
\item{geometry_a, geometry_b}{sfc (simple feature geometry list column)
from \code{\link[=get_geometry]{get_geometry()}}}

\item{x_a, x_b}{X coordinate column, numeric}

\item{y_a, y_b}{Y coordinate column, numeric}

\item{crs}{crs for x_a, y_a (and if provided, x_b, y_b) coordinates,
ignored for geometry_a and geometry_b arguments}
}
\value{
Distance with unit of measurement, see details in \code{\link[sf:geos_measures]{sf::st_distance()}}
}
\description{
\strong{Internal function} - not developed to be used outside of spatsoc functions
}
\details{
Calculate distance using \code{\link[sf:geos_measures]{sf::st_distance()}} for one of the following combinations:
\itemize{
\item the distance matrix of points in geometry_a
\item the distance matrix of points in x_a, y_a
\item the pairwise distance between points in geometry_a and geometry_b
\item the pairwise distance between points in x_a, y_a and x_b, y_b
}

Requirements:
\itemize{
\item matching length between a and b objects if b provided
}
}
\examples{
# Load data.table
library(data.table)
\dontshow{data.table::setDTthreads(1)}

# Example points
example <- data.table(
  X = c(0, 5, 5, 0, 0, NA_real_, 0,        NA_real_),
  Y = c(0, 0, 5, 5, 0, 0,        NA_real_, NA_real_)
)
# E, N, W, S
example[, spatsoc:::calc_distance(x_a = X, y_a = Y, crs = 4326)]
}
\keyword{internal}
