% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/centroids-group.R
\name{CalcCentroids}
\alias{CalcCentroids}
\title{Calculate Centroids For Each Group}
\usage{
CalcCentroids(DT, coordFields, groupField)
}
\arguments{
\item{DT}{Input data.table with coordinate and id fields. If not provided, fields
default to c('EASTING', 'NORTHING') and 'ID'.}

\item{coordFields}{Character vector indicating the X coordinate and Y
coordinate columns. eg: c('EASTING', 'NORTHING')}

\item{groupField}{The group field generated}
}
\description{
Append two columns indicating the centroid coordinates for each group to an
group by ID, coordinates data.table/frame.
}
