% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/direction_polarization.R
\name{direction_group}
\alias{direction_group}
\title{\code{direction_polarization} calculates the direction polarization of
all individuals in each spatiotemporal group identified by \code{group_pts}. The function
expects a \code{data.table} with relocation data appended with a
\code{direction} column from \code{direction_step} and a \code{group} column
from \code{group_pts}.}
\usage{
direction_group(DT, direction = "direction", group = "group")
}
\arguments{
\item{DT}{input data.table with distance column generated by
\code{distance_step} and group column generated with \code{group_pts}}

\item{direction}{character string of direction column name, default
"direction"}

\item{group}{character string of group column name, default "group"}
}
\value{
\code{direction_group} returns the input \code{DT} appended with a
\code{group_direction} column representing the mean direction of
all individuals in each spatiotemporal group.

The mean direction is calculated using \code{\link[CircStats:circ.mean]{CircStats::circ.mean()}}
which expects units of radians.

A message is returned when the \code{group_direction} columns already
exists in the input \code{DT}, because it will be overwritten.
}
\description{
The \code{DT} must be a \code{data.table}. If your data is a
\code{data.frame}, you can convert it by reference using
\code{\link[data.table:setDT]{data.table::setDT}} or by reassigning using
\code{\link[data.table:data.table]{data.table::data.table}}.
}
\details{
The \code{direction} and \code{group} arguments expect the names of columns
in \code{DT} which correspond to the direction and group columns. The
direction column is expected in units of radians and the mean calculated with
\code{\link[CircStats:circ.mean]{CircStats::circ.mean()}}.
}
\examples{
# Load data.table
library(data.table)
\dontshow{data.table::setDTthreads(1)}

# Read example data
DT <- fread(system.file("extdata", "DT.csv", package = "spatsoc"))

# Cast the character column to POSIXct
DT[, datetime := as.POSIXct(datetime, tz = 'UTC')]

# Temporal grouping
group_times(DT, datetime = 'datetime', threshold = '20 minutes')

# Spatial grouping with timegroup
group_pts(DT, threshold = 50, id = 'ID',
          coords = c('X', 'Y'), timegroup = 'timegroup')

# Calculate direction at each step
direction_step(
  DT = DT,
  id = 'ID',
  coords = c('X', 'Y'),
  projection = 32736
)

# Calculate group direction
direction_group(DT)
}
\references{
See example of using mean group direction:
\itemize{
\item \url{https://doi.org/10.1098/rsos.170148}
\item \url{https://doi.org/10.1098/rsos.201128}
\item \url{https://doi.org/10.1016/j.beproc.2018.01.013}
}
}
\seealso{
\code{\link{direction_step}}, \code{\link{group_pts}},
\code{\link[CircStats:circ.mean]{CircStats::circ.mean()}}
}
\concept{Direction functions}
