% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance_to_leader.R
\name{distance_to_leader}
\alias{distance_to_leader}
\title{Distance to group leader}
\usage{
distance_to_leader(
  DT = NULL,
  coords = NULL,
  group = "group",
  crs = NULL,
  geometry = "geometry"
)
}
\arguments{
\item{DT}{input data.table with 'rank_position_group_direction' column
generated by \code{leader_direction_group} and group column generated by
\code{group_pts}}

\item{coords}{character vector of X coordinate and Y coordinate column names.
Note: the order is assumed X followed by Y column names.}

\item{group}{group column name, generated by \code{group_pts}, default
'group'}

\item{crs}{numeric or character defining the coordinate reference
system to be passed to \link[sf:st_crs]{sf::st_crs}. For example, either
\code{crs = "EPSG:32736"} or \code{crs = 32736}. If \code{crs = NULL}, the \code{crs} will be
internally set to \code{sf::NA_crs_}.}
}
\value{
\code{distance_to_leader} returns the input \code{DT} appended with a
\code{distance_leader} column indicating the distance to the group leader.

A message is returned when the \code{distance_leader} column already exist in
the input \code{DT} because it will be overwritten.

See details for appending outputs using modify-by-reference in the
\href{https://docs.ropensci.org/spatsoc/articles/faq.html}{FAQ}.
}
\description{
\code{distance_to_leader} calculates the distance to the leader of each
spatiotemporal group. The function expects a \code{data.table} with relocation
data appended with a \code{rank_position_group_direction} column indicating the
ranked position along the group direction generated with
\code{leader_direction_group(return_rank = TRUE)}. Relocation data should be in
two columns representing the X and Y coordinates, or in a geometry column
prepared by the helper function \code{\link[=get_geometry]{get_geometry()}}.
}
\details{
The \code{DT} must be a \code{data.table}. If your data is a \code{data.frame}, you can
convert it by reference using \code{\link[data.table:setDT]{data.table::setDT()}} or by reassigning using
\code{\link[data.table:data.table]{data.table::data.table()}}.

This function expects a \code{rank_position_group_direction} column generated with
\code{leader_direction_group(return_rank = TRUE)}, a \code{group} column generated with
the \code{group_pts} function. The \code{group} argument expect the names of the column
in \code{DT} which corresponds to the group column.

See below under "Interface" for details on providing coordinates and under
"Distance function" for details on underlying distance function used.
}
\section{Interface}{

Two interfaces are available for providing coordinates:
\enumerate{
\item Provide \code{coords} and \code{crs}. The \code{coords} argument expects the names of
the X and Y coordinate columns. The \code{crs} argument expects a character
string or numeric defining the coordinate reference system to be passed to
\link[sf:st_crs]{sf::st_crs}. For example, for UTM zone 36S (EPSG 32736), the crs argument
is \code{crs = "EPSG:32736"} or \code{crs = 32736}. See \url{https://spatialreference.org}
for a list of EPSG codes.
\item (New!) Provide \code{geometry}. The \code{geometry} argument allows the user to
supply a \code{geometry} column that represents the coordinates as a simple
feature geometry list column. This interface expects the user to prepare
their input DT with \code{\link[=get_geometry]{get_geometry()}}. To use this interface, leave the
\code{coords} and \code{crs} arguments \code{NULL}, and the default argument for \code{geometry}
('geometry') will be used directly.
}
}

\section{Distance function}{


The underlying distance function used depends on the crs of the coordinates
or geometry provided.
\itemize{
\item If the crs is longlat degrees (as determined by
\code{\link[sf:st_is_longlat]{sf::st_is_longlat()}}), the distance function is \code{\link[sf:geos_measures]{sf::st_distance()}} which
passes to \code{\link[s2:s2_is_collection]{s2::s2_distance()}} if \code{\link[sf:s2]{sf::sf_use_s2()}} is TRUE and
\code{\link[lwgeom:geod]{lwgeom::st_geod_distance()}} if \code{\link[sf:s2]{sf::sf_use_s2()}} is FALSE. The distance
returned has units set according to the crs.
\item If the crs is not longlat degrees (eg. NULL, NA_crs_, or projected), the
distance function used is \code{\link[stats:dist]{stats::dist()}}, maintaining expected behaviour
from previous versions. The distance returned does not have units set.
}

Note: in both cases, if the coordinates are NA then the result will be NA.
}

\examples{
# Load data.table
library(data.table)
\dontshow{data.table::setDTthreads(1)}

# Read example data
DT <- fread(system.file("extdata", "DT.csv", package = "spatsoc"))

# Cast the character column to POSIXct
DT[, datetime := as.POSIXct(datetime, tz = 'UTC')]

# (Subset example data to reduce example run time)
DT <- DT[year(datetime) == 2016]

# Temporal grouping
group_times(DT, datetime = 'datetime', threshold = '20 minutes')

# Spatial grouping with timegroup
group_pts(DT, threshold = 50, id = 'ID',
          coords = c('X', 'Y'), timegroup = 'timegroup')

# Calculate direction at each step
direction_step(
  DT = DT,
  id = 'ID',
  coords = c('X', 'Y'),
  crs = 32736
)

# Calculate group centroid
centroid_group(DT, coords = c('X', 'Y'))

# Calculate group direction
direction_group(DT)

# Calculate leader in terms of position along group direction
leader_direction_group(
  DT,
  coords = c('X', 'Y'),
  return_rank = TRUE
)

# Calculate distance to leader
distance_to_leader(DT, coords = c('X', 'Y'), crs = 32736)
}
\references{
See examples of using distance to leader and position within group:
\itemize{
\item \doi{doi:10.1111/jfb.15315}
\item \doi{doi:10.1098/rspb.2017.2629}
\item \doi{doi:10.1016/j.anbehav.2023.09.009}
}
}
\seealso{
\link{direction_to_leader}, \link{leader_direction_group}, \link{group_pts},
\code{\link[sf:geos_measures]{sf::st_distance()}}

Other Distance functions: 
\code{\link{distance_to_centroid}()},
\code{\link{edge_nn}()},
\code{\link{edge_zones}()}
}
\concept{Distance functions}
