% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_gbi.R
\name{get_gbi}
\alias{get_gbi}
\title{Generate group by individual matrix}
\usage{
get_gbi(DT = NULL, group = "group", id = NULL, type = NULL)
}
\description{
inheritParams group_pts
param
}
\seealso{
\code{\link{group_pts}} \code{\link{group_lines}} \code{\link{group_polys}}
family
importFrom
examples
}
