% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_lines.R
\name{group_lines}
\alias{group_lines}
\title{Groups From Spatial Lines}
\usage{
group_lines(DT = NULL, threshold = NULL, projection = NULL, id = NULL,
  coordFields = NULL, timegroup = NULL, groupFields = NULL,
  spLines = NULL)
}
\arguments{
\item{DT}{input data.table}

\item{threshold}{The width of the buffer around the geometry in the units
of the projection. Optionally, exclude argument or supply 0 to compare
SpatialLines intersection}

\item{projection}{PROJ.4 charaster string}

\item{id}{Character string of ID column name}

\item{coordFields}{Character vector of X coordinate and Y coordinate column names}

\item{timegroup}{(optional) timegroup field in the DT upon which the grouping will be calculated}

\item{groupFields}{(optional) character string or vector of grouping field(s) upon which the grouping will be calculated}

\item{spLines}{Alternatively, provide a SpatialLines object created with the sp
package. If a spLines object is provided, groups cannot be calculated by a
timegroup}
}
\value{
Group by ID (by time) data.table
}
\description{
Groups From Spatial Lines
}
\examples{
utm <- '+proj=utm +zone=36 +south +ellps=WGS84 +datum=WGS84 +units=m +no_defs'


group_lines(locs, threshold = 50, projection = utm,
            id = 'ID', coordFields = c('X', 'Y'))

# Daily movement tracks
group_times(locs, timeField = 'datetime', threshold = '1 day')
group_lines(locs, threshold = 50, projection = utm,
            id = 'ID', coordFields = c('X', 'Y'),
            timegroup = 'timegroup')

# Daily movement tracks by sex
group_times(locs, timeField = 'datetime', threshold = '1 day')
group_lines(locs, threshold = 50, projection = utm,
            id = 'ID', coordFields = c('X', 'Y'),
            timegroup = 'timegroup', groupFields = 'sex')
}
