% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_polys.R
\name{group_polys}
\alias{group_polys}
\title{Group Overlapping Polygons}
\usage{
group_polys(DT = NULL, area = NULL, hrType = NULL, hrParams = NULL,
  projection = NULL, coordFields = NULL, idField = NULL,
  groupFields = NULL, spPolys = NULL)
}
\arguments{
\item{DT}{input data.table}

\item{area}{boolean indicating either returning area and proportion of overlap or group}

\item{hrType}{type of HR estimation, of 'mcp' or 'kernel'}

\item{hrParams}{parameters for adehabitatHR functions, a named list passed to do.call}

\item{projection}{PROJ.4 charaster string}

\item{coordFields}{Character vector of X coordinate and Y coordinate column names}

\item{idField}{Character string of ID column name}

\item{groupFields}{(optional) character string or vector of grouping field(s) upon which the grouping will be calculated}

\item{spPolys}{Alternatively, provide solely a SpatialPolygons object.}
}
\value{
Group by ID (by time) data.table
}
\description{
Group individuals by polygon (eg: home ranges) overlap
}
\examples{
groups <- group_polys(locs, area = FALSE, 'mcp', list(percent = 95),
                      projection = utm,
                      idField = 'ID', coordFields = c('X', 'Y'))

areaDT <- group_polys(locs, area = TRUE, 'mcp', list(percent = 95),
                      projection = utm,
                      idField = 'ID', coordFields = c('X', 'Y'))
}
