% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group-lines.R
\name{GroupLines}
\alias{GroupLines}
\title{Groups From Spatial Lines}
\usage{
GroupLines(DT = NULL, threshold = NULL, projection = NULL,
  idField = NULL, coordFields = NULL, timeGroup = NULL,
  groupFields = NULL, spLines = NULL)
}
\arguments{
\item{DT}{input data.table}

\item{threshold}{The width of the buffer around the geometry in the units
of the projection. Optionally, exclude argument or supply 0 to compare
SpatialLines intersection}

\item{projection}{PROJ.4 charaster string}

\item{idField}{Character string of ID column name}

\item{coordFields}{Character vector of X coordinate and Y coordinate column names}

\item{timeGroup}{(optional) timegroup field in the DT upon which the grouping will be calculated}

\item{groupFields}{(optional) character string or vector of grouping field(s) upon which the grouping will be calculated}

\item{spLines}{Alternatively, provide a SpatialLines object created with the sp
package. If a spLines object is provided, groups cannot be calculated by a
timeGroup}
}
\value{
Group by ID (by time) data.table
}
\description{
Groups From Spatial Lines
}
\examples{
data(locs)
groups <- GroupLines(locs, 50)

groups <- GroupLines(locs, 50, timeGroup = 'FIX_DATE',
        projection = '+proj=utm +zone=21 ellps=WGS84',
        idField = 'ID')


data(locsLines)

groups <- GroupLines(spLines = locsLines)
}
