% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group-pts.R
\name{GroupPts}
\alias{GroupPts}
\title{Group Points By Buffer}
\usage{
GroupPts(DT, distance, time = NULL, groupFields = NULL,
  coordFields = c("EASTING", "NORTHING"), idField = "ID")
}
\arguments{
\item{DT}{Input data with coordinate and id fields. If not provided, fields
default to c('EASTING', 'NORTHING') and 'ID'}

\item{distance}{The threshold distance for grouping points. The distance must be in the units of the projection.}

\item{time}{(optional) time field in the DT upon which the spatial grouping will be calculated.}

\item{groupFields}{(optional) grouping field(s) to be combined with the time field, either character or list of strings.}

\item{coordFields}{Character vector indicating the X coordinate and Y
coordinate columns. eg: c('EASTING', 'NORTHING')}

\item{idField}{Character string indicating the column name for the ID
associated with each point}
}
\value{
Input DT with column "group" added.
}
\description{
Group points by buffer overlap at equal time.
}
\details{
This function uses input spatial points to determine groups in space and
time.
}
\examples{

GroupPts(locs, distance = 50)

GroupPts(locs, distance = 50, time = 'datetime')

GroupPts(locs, distance = 50, time = 'timegroup')

GroupPts(locs, distance = 50, time = 'timegroup', groupFields = 'season')
}
