% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group-pts.R
\name{GroupPts}
\alias{GroupPts}
\title{Group Points By Buffer}
\usage{
GroupPts(DT = NULL, distance = NULL, timeGroup = NULL,
  groupFields = NULL, coordFields = NULL, idField = NULL)
}
\arguments{
\item{DT}{Input data.table with coordinate and id fields. If not provided, fields
default to c('EASTING', 'NORTHING') and 'ID'.}

\item{distance}{The threshold distance for grouping points. The distance must be in the units of the projection.}

\item{timeGroup}{(optional) time group field in the DT upon which the spatial grouping will be calculated.}

\item{groupFields}{(optional) grouping field(s) to be combined with the timeGroup field, either character or list of strings.}

\item{coordFields}{Character vector indicating the X coordinate and Y
coordinate columns. eg: c('EASTING', 'NORTHING')}

\item{idField}{Character string indicating the column name for the ID
associated with each point.}
}
\value{
Input DT with column 'group' added.
}
\description{
Group points by spatial distance and temporal overlap.
}
\details{
This function finds spatialtemporal groups in input points.
}
\examples{

GroupPts(locs, distance = 50)

GroupPts(locs, distance = 50, timeGroup = 'datetime')

GroupPts(locs, distance = 50, timeGroup = 'timegroup')

GroupPts(locs, distance = 50, timeGroup = 'timegroup', groupFields = 'season')
}
