% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group-pts.R
\name{GroupPts}
\alias{GroupPts}
\title{Group Points By Buffer}
\usage{
GroupPts(dt, bufferWidth, timeField = NULL, crs, coordFields = c("EASTING",
  "NORTHING"), idField = "ID", spPts = NULL)
}
\arguments{
\item{dt}{Input data with coordinate and id fields. If not provided, fields default to c('EASTING', 'NORTHING') and 'ID'}

\item{bufferWidth}{The width of the buffer around the geometry in the units of the projection}

\item{timeField}{Time field in the dt upon which the spatial grouping will be
calculated}

\item{crs}{Character string for input to sp::CRS()}

\item{coordFields}{Character vector indicating the X coordinate and Y
coordinate columns. eg: c('EASTING', 'NORTHING')}

\item{idField}{Character string indicating the column name for the ID associated
with each point}

\item{spPts}{Alternatively, provide a SpatialPointsDataFrame created with the
sp package. If a spPts object is provided, groups cannot be calculated by}
}
\description{
Group points by buffer overlap at equal time.
}
\details{
This function uses input spatial points to determine groups in space and
time.
}
