% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group-time.R
\name{GroupTimes}
\alias{GroupTimes}
\title{Group Times}
\usage{
GroupTimes(DT = NULL, timeField = NULL, threshold = NULL)
}
\arguments{
\item{DT}{input locs/rows}

\item{timeField}{time column name}

\item{threshold}{character defining the threshold for grouping times.
eg: '2 hours', '10 minutes', etc.
if not provided, times will be matched exactly.
Note that provided threshold must be in the expected format: '## unit'}
}
\description{
Assign an integer group for rows by provided time column. If
using GPS collar data or other data with variability in concurrent measures
across individuals, it is usefuls to round this time column
providing the threshold.
Otherwise, rows are grouped by matching exact times.
}
\details{
This function can also group rows on blocks of days.
Simply provide the block length in the same manner as the threshold eg: '5 days'.
}
