% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group-time.R
\name{GroupTimes}
\alias{GroupTimes}
\title{Group Times}
\usage{
GroupTimes(DT, timeField, timeThreshold = NULL)
}
\arguments{
\item{DT}{input locs/rows}

\item{timeField}{time column name}

\item{timeThreshold}{character defining the threshold for grouping times.
eg: '2 hours', '10 minutes', etc.
if not provided, times will be matched exactly.
Note that provided threshold must be in the expected format: '## unit'}
}
\description{
Assign a numerical group for rows by provided time column. If
using GPS collar data or other data with variability in concurrent measures
across individuals, it is recommended to round this time column
providing the timeThreshold.
Otherwise, rows are grouped by matching exact times.
}
\details{
This function can also group rows on time intervals, such as blocks of 5 days.
Simply provide the interval in the same manner eg: '5 days'.
}
