% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group-lines.R
\name{grp_lines}
\alias{grp_lines}
\title{Groups From Spatial Lines}
\usage{
grp_lines(dt, buffer.width, crs, coord.fields = c("EASTING", "NORTHING"),
  id.field = "ID", sp.lines = NULL)
}
\arguments{
\item{dt}{Input data with coordinate and id fields. If not provided, fields default to c('EASTING', 'NORTHING') and 'ID'}

\item{buffer.width}{The width of the buffer around the SpatialLines in the
units of the projection provided.}

\item{crs}{Character string for input to sp::CRS()}

\item{coord.fields}{Character vector indicating the X coordinate and Y
coordinate columns. eg: c('EASTING', 'NORTHING')}

\item{id.field}{Character string indicating the column name for the ID associated
with each point}
}
\value{
id by spatial group data.table
}
\description{
Groups From Spatial Lines
}
\examples{

}
