% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group-pts.R
\name{grp_pts}
\alias{grp_pts}
\title{Group Points By Buffer}
\usage{
grp_pts(dt, buffer.width, crs, coord.fields = c("EASTING", "NORTHING"),
  id.field = "ID", sp.pts = NULL)
}
\arguments{
\item{dt}{Input data with coordinate and id fields. If not provided, fields default to c('EASTING', 'NORTHING') and 'ID'}

\item{buffer.width}{The width of the buffer in the units of the projection}

\item{crs}{Character string for input to sp::CRS()}

\item{coord.fields}{Character vector indicating the X coordinate and Y
coordinate columns. eg: c('EASTING', 'NORTHING')}

\item{id.field}{Character string indicating the column name for the ID associated
with each point}

\item{sp.pts}{Alternatively, provide a SpatialPointsDataFrame created with the sp package.}
}
\description{
Group points by buffer overlap at each time interval
}
\details{
this function uses input spatial points to determine
}
