% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nearest.R
\name{Nearest}
\alias{Nearest}
\title{Proportion/Frequency Nearest Neighbors}
\usage{
Nearest(DT, timeField = NULL, groupField = NULL, proportions = FALSE,
  coordFields = c("EASTING", "NORTHING"), idField = "ID")
}
\arguments{
\item{DT}{Input data.table with coordinate and id fields. If not provided, fields
default to c('EASTING', 'NORTHING') and 'ID'.}

\item{timeField}{(optional) timeField in the DT upon which the neighbors'
distance will be compared}

\item{groupField}{(optional) groupField in the DT which can be used to group
neighbors (eg: season, year, herd, known social groups, ...)}

\item{coordFields}{Character vector indicating the X coordinate and Y
coordinate columns. eg: c('EASTING', 'NORTHING')}

\item{idField}{Character string indicating the column name for the ID
associated with each point.}
}
\description{
Proportion/Frequency Nearest Neighbors
}
