% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nearest.R
\name{Nearest}
\alias{Nearest}
\title{Proportion/Frequency Nearest Neighbors}
\usage{
Nearest(dt, timeField = NULL, coordFields = c("EASTING", "NORTHING"),
  idField = "ID")
}
\arguments{
\item{dt}{Input data with coordinate and id fields. If not provided, fields default to c('EASTING', 'NORTHING') and 'ID'}

\item{timeField}{Time field in the dt upon which the nearest neighbors will
be calculate}

\item{coordFields}{Character vector indicating the X coordinate and Y
coordinate columns. eg: c('EASTING', 'NORTHING')}

\item{idField}{Character string indicating the column name for the ID associated
with each point}
}
\description{
Proportion/Frequency Nearest Neighbors
}
