% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nearest.R
\name{Nearest}
\alias{Nearest}
\title{Proportion/Frequency Nearest Neighbors}
\usage{
Nearest(DT, timeField = NULL, groupField = NULL, proportions = FALSE,
  coordFields = c("EASTING", "NORTHING"), id = "ID")
}
\arguments{
\item{DT}{input data.table}

\item{timeField}{(optional) timeField in the DT upon which the neighbors'
distance will be compared}

\item{groupField}{(optional) groupField in the DT which can be used to group
neighbors (eg: season, year, herd, known social groups, ...)}

\item{coordFields}{Character vector of X coordinate and Y coordinate column names}

\item{id}{Character string of ID column name}
}
\description{
Proportion/Frequency Nearest Neighbors
}
