% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean-pairwise-dist.R
\name{PairwiseDist}
\alias{PairwiseDist}
\title{Euclidean Pairwise Distance}
\usage{
PairwiseDist(DT, timeField, coordFields = c("EASTING", "NORTHING"),
  idField = "ID")
}
\arguments{
\item{DT}{input data.table}

\item{timeField}{time column upon which individuals are compared. recommended
to use a rounded time, see SOME LINK.}

\item{coordFields}{Character vector of X coordinate and Y coordinate column names}

\item{idField}{Character string of ID column name}
}
\value{
mean pairwise distance data.table by ID and (optional) time field
}
\description{
Euclidean Pairwise Distance
}
