% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean-pairwise-dist.R
\name{PairwiseDist}
\alias{PairwiseDist}
\title{Euclidean Pairwise Distance}
\usage{
PairwiseDist(dt, timeField, coordFields = c("EASTING", "NORTHING"),
  idField = "ID")
}
\arguments{
\item{dt}{Input data with coordinate and id fields. If not provided, fields default to c('EASTING', 'NORTHING') and 'ID'}

\item{timeField}{time column upon which individuals are compared. recommended
to use a rounded time, see SOME LINK.}

\item{coordFields}{Character vector indicating the X coordinate and Y
coordinate columns. eg: c('EASTING', 'NORTHING')}

\item{idField}{Character string indicating the column name for the ID associated
with each point}
}
\value{
mean pairwise distance data.table by ID and (optional) time field
}
\description{
Euclidean Pairwise Distance
}
