% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomizations.R
\name{Randomizations}
\alias{Randomizations}
\title{Randomizations}
\usage{
Randomizations(DT = NULL, type = NULL, id = NULL, dateField = NULL,
  splitBy = NULL, iterations = NULL)
}
\arguments{
\item{DT}{input data.table with id, group fields and (optional) time fields}

\item{type}{one of 'daily', 'hourly' or 'spiegel' - see details}

\item{id}{field indicating the id in the input data.table}

\item{splitBy}{List of fields in DT to split the randomization process by}

\item{iterations}{The number of iterations to randomize}

\item{groupField}{field indicating the group membership for each id}

\item{timeField}{(optional) time field used for providing datetime or hour field or group time field}
}
\description{
Randomization
}
\details{
Randomization types:

'hourly' simply randomly assigns an ID to each location.

'daily' will randomize the ID for each individual 24hr trajectory

'spiegel' will implement the daily movement trajectory randomizations (Spiegel et al. 2016).
}
\seealso{
\url{http://onlinelibrary.wiley.com/doi/10.1111/2041-210X.12553/full}
}
