% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomizations.R
\name{randomizations}
\alias{randomizations}
\title{Randomizations}
\usage{
randomizations(DT = NULL, type = NULL, id = NULL, datetime = NULL,
  splitBy = NULL, iterations = NULL)
}
\arguments{
\item{DT}{input data.table with id, group fields and (optional) time fields}

\item{type}{one of 'daily', 'step' or 'trajectory' - see details}

\item{id}{field indicating the id in the input data.table}

\item{datetime}{(optional) time field used for providing datetime or hour field or group time field}

\item{splitBy}{List of fields in DT to split the randomization process by}

\item{iterations}{The number of iterations to randomize}
}
\description{
Data stream randomization methods
}
\details{
Three randomization methods are provided:

'step' randomly assigns an ID to each location.

'daily' will randomize the ID for each individual 24hr trajectory

'trajectory' will implement the daily movement trajectory randomizations (Spiegel et al. 2016).
}
\references{
\url{http://onlinelibrary.wiley.com/doi/10.1111/2041-210X.12553/full}
}
