% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatsoc.R
\docType{package}
\name{spatsoc}
\alias{spatsoc}
\alias{_PACKAGE}
\alias{spatsoc-package}
\title{spatsoc}
\description{
spatsoc is an R package for detecting spatial and temporal groups in GPS relocations. It can be used to convert GPS relocations to gambit-of-the-group format to build proximity-based social networks. In addition, the randomization function provides data-stream randomization methods suitable for GPS data.
}
\details{
The spatsoc package provides one temporal grouping function:

\itemize{
\item \code{\link{group_times}}
}
three spatial grouping functions:
\itemize{
\item \code{\link{group_pts}}
\item \code{\link{group_lines}}
\item \code{\link{group_polys}}
}

and two social network functions:
\itemize{
\item \code{\link{randomizations}}
\item \code{\link{get_gbi}}
}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://spatsoc.gitlab.io}
  \item \url{https://www.gitlab.com/robit.a/spatsoc}
  \item \url{https://github.com/ropensci/spatsoc}
  \item Report bugs at \url{https://www.gitlab.com/robit.a/spatsoc/issues}
}

}
\author{
\strong{Maintainer}: Alec L. Robitaille \email{robit.alec@gmail.com}

Authors:
\itemize{
  \item Quinn Webber
  \item Eric Vander Wal
}

}
